models: add collection_ref as child model
collection_ref is useful for cases which models that belongs to
different collections, must be aggregated into one single collection
that basically have only the model ids and a refference for its
collections. Example: in this application we have multiple devices
collections (e.g. sonoffs, tvs, acs, rgb lights, etc). To retrieve a
device, I have to specify its id and the collection that the device belongs to.
But in some cases, it would be desirable to retrieve a device only
knowing its id, or retrieve all the existing devices. In these
situations, I can just query the collection that stores the collection_ref
children to get all devices or to get the collection of a device by its
id.